refactor : shared IconButton을 사용하도록 리팩토링
