+++ src/features/ContentOverviewHero/ui/ContentOverviewActionButtons.tsx
@@ -1,4 +1,5 @@
 import { MapPin, Map } from 'lucide-react';
+import { IconButton } from '@/shared/ui/IconButton';
 import type { ContentOverviewActionButtonsProps } from '../model/types';
 import { messages } from '../model/messages';
 
@@ -10,26 +11,25 @@ export function ContentOverviewActionButtons({
     <div className="absolute bottom-8 left-0 right-0 z-10 px-6">
       <div className="max-w-2xl mx-auto">
         <div className="flex flex-col sm:flex-row gap-4 justify-center">
-          <button
+          <IconButton
+            Icon={MapPin}
+            shape="pill"
+            size="lg"
+            className="bg-white text-black font-semibold px-8 py-4 hover:bg-gray-50"
             onClick={onLocationViewClick}
-            className="
-              flex items-center justify-center gap-3 px-8 py-4 
-              bg-white font-semibold rounded-xl"
           >
-            <MapPin className="w-5 h-5" />
             {messages.locationView}
-          </button>
+          </IconButton>
 
-          <button
+          <IconButton
+            Icon={Map}
+            shape="pill"
+            size="lg"
+            className="bg-white/15 text-white font-semibold px-8 py-4 border border-white/30 hover:bg-white/25"
             onClick={onMapViewClick}
-            className="
-              flex items-center justify-center gap-3 px-8 py-4 
-              bg-white/15 text-white font-semibold rounded-xl 
-              border border-white/30"
           >
-            <Map className="w-5 h-5" />
             {messages.mapView}
-          </button>
+          </IconButton>
         </div>
       </div>
     </div>
