+++ src/features/Header/ui/ProfileButton.tsx
@@ -1,5 +1,5 @@
 import { User } from 'lucide-react';
-import { IconButton } from '@/shared/ui/IconButton';
+import { IconButton } from '@/shared/ui';
 
 export function ProfileButton() {
   return <IconButton Icon={User} variant="gradient" shape="circle" size="md" aria-label="프로필" />;
